cloud_function_directory/requirements.txt
google-cloud-storage==1.42.3
requests==2.26.0
google-auth==2.3.3